+++ crates/api/src/server.rs
@@ -43,7 +43,8 @@ pub async fn serve(
 ) {
     let rib_svc = RibService::new(rib_tx.clone());
     let injection_svc = InjectionService::new(rib_tx.clone());
-    let neighbor_svc = NeighborService::new(peer_mgr_tx.clone(), rib_tx.clone(), config_tx);
+    let neighbor_svc =
+        NeighborService::new(config.asn, peer_mgr_tx.clone(), rib_tx.clone(), config_tx);
     let global_svc = GlobalService::new(config.asn, config.router_id, config.listen_port);
     let control_svc = ControlService::new(
         config.start_time,